pyproject.toml
[project]
name = "lab-3"
version = "0.1.0"
description = "Quantum Computing Lab 2 Exercises"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyfunctional>=1.5.0",
    "bitarray>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
